s1c4 - gah.
Took a long time and a diversion because I failed to dehex
the lines first! Current impl of 'english_score' is simple
tiered counting. I thought the more complex histogram distance
(chi-squared) and/or ordering distance didn't work because of
the failed de-hexing.

If we need a better english discriminator in the future, dig those
out again.